Move model tests into unit tests
Ref. #252
